Fix compilation error with Magick.NET 14.10.2 API change
Cast JpegQuality to uint for Magick.NET 14.10.2 API compatibility.
The Quality property now requires uint instead of int.

+++ src/Core/ConversionEngine.cs
@@ -59,7 +59,7 @@ public class ConversionEngine
 
                 using var image = new MagickImage(sourcePath);
                 image.Format = MagickFormat.Jpeg;
-                image.Quality = _settings.JpegQuality;
+                image.Quality = (uint)_settings.JpegQuality;
 
                 // Preserve EXIF orientation
                 image.AutoOrient();
